Port public breach index to proper React
