+++ src/app/(proper_react)/(redesign)/(public)/breaches/BreachIndexView.module.scss
@@ -1,0 +1,93 @@
+@import "../../../../tokens";
+
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: $spacing-lg;
+  padding: $spacing-lg;
+
+  header {
+    padding: $spacing-2xl $spacing-xl;
+    max-width: $content-lg;
+    text-align: center;
+    display: flex;
+    flex-direction: column;
+    gap: $spacing-md;
+
+    h1 {
+      font: $text-title-sm;
+    }
+  }
+}
+
+.breachList {
+  width: $content-xl;
+  max-width: calc(100% - 2 * $spacing-2xl);
+  display: grid;
+  align-items: start;
+  grid-gap: $spacing-2xl;
+  grid-template-columns: repeat(auto-fit, minmax($content-xs, 1fr));
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+
+  li {
+    height: 100%;
+  }
+
+  .breachCard {
+    display: flex;
+    flex-direction: column;
+    box-shadow: $box-shadow-sm;
+    border-radius: $border-radius-md;
+    height: 100%;
+    text-decoration: none;
+    color: inherit;
+
+    &:hover {
+      box-shadow: $box-shadow-md;
+    }
+
+    h2 {
+      background-color: $color-grey-05;
+      display: flex;
+      align-items: center;
+      gap: $spacing-md;
+      padding: $spacing-md $spacing-lg;
+    }
+
+    dl {
+      flex-grow: 1;
+      display: flex;
+      flex-direction: column;
+      gap: $spacing-md;
+      padding: $spacing-lg $spacing-md;
+      // `.breachCard h2` padding + breach logo width + `.breachCard h2` gap,
+      // to align the content to the title in the header:
+      padding-inline-start: calc($spacing-md + 32px + $spacing-md);
+
+      dt {
+        font: $text-body-sm;
+        color: $color-grey-50;
+      }
+
+      dd {
+        font-weight: 700;
+      }
+    }
+
+    .linkDescription {
+      padding: $spacing-md;
+      // `.breachCard h2` padding + breach logo width + `.breachCard h2` gap,
+      // to align the content to the title in the header:
+      padding-inline-start: calc($spacing-md + 32px + $spacing-md);
+      color: $color-blue-50;
+      font-weight: 700;
+
+      &:hover {
+        text-decoration: underline;
+      }
+    }
+  }
+}
